pyproject.toml
[project]
name = "ultrawide-window-positioner"
version = "1.3.0.0"
description = "Application to manage window position, size and always-on-top status."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    'global-hotkeys>=0.1.7 ; sys_platform == "win32"',
    'hdrcapture>=0.3.0 ; sys_platform == "win32"',
    "psutil>=7.2.2",
    "pyside6-essentials>=6.11.0",
    "roman>=5.2",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.19.0",
]

[tool.uv.sources]
hdrcapture = { path = "hdrcapture-0.3.0-cp312-cp312-win_amd64.whl", marker = "sys_platform == 'win32'" }
